+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ba/XFactory.java
@@ -227,18 +227,6 @@ public class XFactory {
 
     }
 
-    /**
-     * @deprecated This method does not work as expected. Will be removed in 5.x release.
-     */
-    @Deprecated
-    public boolean nameAndSignatureIsCalled(XMethod m) {
-        return calledMethodSignatures.contains(getDetailedSignature(m));
-    }
-
-    private static String getDetailedSignature(XMethod m2) {
-        return m2.getName() + m2.getSignature() + m2.isStatic();
-    }
-
     @Deprecated
     public boolean isInterned(XMethod m) {
         return m.isResolved();